updated colorize
